Crate rusoto_opsworkscm

AWS OpsWorks CM
AWS OpsWorks for configuration management (CM) is a service that runs and manages configuration management servers. You can use AWS OpsWorks CM to create and manage AWS OpsWorks for Chef Automate and AWS OpsWorks for Puppet Enterprise servers, and add or remove nodes for the servers to manage.

Glossary of terms

Server: A configuration management server that can be highly-available. The configuration management server runs on an Amazon Elastic Compute Cloud (EC2) instance, and may use various other AWS services, such as Amazon Relational Database Service (RDS) and Elastic Load Balancing. A server is a generic abstraction over the configuration manager that you want to use, much like Amazon RDS. In AWS OpsWorks CM, you do not start or stop servers. After you create servers, they continue to run until they are deleted.

Engine: The engine is the specific configuration manager that you want to use. Valid values in this release include ChefAutomate and Puppet.

Backup: This is an application-level backup of the data that the configuration manager stores. AWS OpsWorks CM creates an S3 bucket for backups when you launch the first server. A backup maintains a snapshot of a server's configuration-related attributes at the time the backup starts.

Events: Events are always related to a server. Events are written during server creation, when health checks run, when backups are created, when system maintenance is performed, etc. When you delete a server, the server's events are also deleted.

Account attributes: Every account has attributes that are assigned in the AWS OpsWorks CM database. These attributes store information about configuration limits (servers, backups, etc.) and your customer account.

Endpoints

AWS OpsWorks CM supports the following endpoints, all HTTPS. You must connect to one of the following endpoints. Your servers can only be accessed or managed within the endpoint in which they are created.

opsworks-cm.us-east-1.amazonaws.com

opsworks-cm.us-east-2.amazonaws.com

opsworks-cm.us-west-1.amazonaws.com

opsworks-cm.us-west-2.amazonaws.com

opsworks-cm.ap-northeast-1.amazonaws.com

opsworks-cm.ap-southeast-1.amazonaws.com

opsworks-cm.ap-southeast-2.amazonaws.com

opsworks-cm.eu-central-1.amazonaws.com

opsworks-cm.eu-west-1.amazonaws.com

Throttling limits

All API operations allow for five requests per second with a burst of 10 requests per second.

If you're using the service, you're probably looking for OpsWorksCMClient and OpsWorksCM.

Structs

AccountAttribute
Stores account attributes.

AssociateNodeRequest

AssociateNodeResponse

Backup
Describes a single backup.

CreateBackupRequest

CreateBackupResponse

CreateServerRequest

CreateServerResponse

DeleteBackupRequest

DeleteBackupResponse

DeleteServerRequest

DeleteServerResponse

DescribeAccountAttributesRequest

DescribeAccountAttributesResponse

DescribeBackupsRequest

DescribeBackupsResponse

DescribeEventsRequest

DescribeEventsResponse

DescribeNodeAssociationStatusRequest

DescribeNodeAssociationStatusResponse

DescribeServersRequest

DescribeServersResponse

DisassociateNodeRequest

DisassociateNodeResponse

EngineAttribute
A name and value pair that is specific to the engine of the server.

ExportServerEngineAttributeRequest

ExportServerEngineAttributeResponse

OpsWorksCMClient
A client for the OpsWorksCM API.

RestoreServerRequest

RestoreServerResponse

Server
Describes a configuration management server.

ServerEvent
An event that is related to the server, such as the start of maintenance or backup.

StartMaintenanceRequest

StartMaintenanceResponse

UpdateServerEngineAttributesRequest

UpdateServerEngineAttributesResponse

UpdateServerRequest

UpdateServerResponse

Enums

AssociateNodeError
Errors returned by AssociateNode

CreateBackupError
Errors returned by CreateBackup

CreateServerError
Errors returned by CreateServer

DeleteBackupError
Errors returned by DeleteBackup

DeleteServerError
Errors returned by DeleteServer

DescribeAccountAttributesError
Errors returned by DescribeAccountAttributes

DescribeBackupsError
Errors returned by DescribeBackups

DescribeEventsError
Errors returned by DescribeEvents

DescribeNodeAssociationStatusError
Errors returned by DescribeNodeAssociationStatus

DescribeServersError
Errors returned by DescribeServers

DisassociateNodeError
Errors returned by DisassociateNode

ExportServerEngineAttributeError
Errors returned by ExportServerEngineAttribute

RestoreServerError
Errors returned by RestoreServer

StartMaintenanceError
Errors returned by StartMaintenance

UpdateServerEngineAttributesError
Errors returned by UpdateServerEngineAttributes

UpdateServerError
Errors returned by UpdateServer

Traits

OpsWorksCM
Trait representing the capabilities of the OpsWorksCM API. OpsWorksCM clients implement this trait.