Struct rusoto_opsworkscm::Backup

pub struct Backup {
    pub backup_arn: Option<String>,
    pub backup_id: Option<String>,
    pub backup_type: Option<String>,
    pub created_at: Option<f64>,
    pub description: Option<String>,
    pub engine: Option<String>,
    pub engine_model: Option<String>,
    pub engine_version: Option<String>,
    pub instance_profile_arn: Option<String>,
    pub instance_type: Option<String>,
    pub key_pair: Option<String>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub s3_log_url: Option<String>,
    pub security_group_ids: Option<Vec<String>>,
    pub server_name: Option<String>,
    pub service_role_arn: Option<String>,
    pub status: Option<String>,
    pub status_description: Option<String>,
    pub subnet_ids: Option<Vec<String>>,
    pub tools_version: Option<String>,
    pub user_arn: Option<String>,
}

Describes a single backup.

Fields
backup_arn: Option<String>

The ARN of the backup.

backup_id: Option<String>

The generated ID of the backup. Example: myServerName-yyyyMMddHHmmssSSS

backup_type: Option<String>

The backup type. Valid values are automated or manual.

created_at: Option<f64>

The time stamp when the backup was created in the database. Example: 2016-07-29T13:38:47.520Z

description: Option<String>

A user-provided description for a manual backup. This field is empty for automated backups.

engine: Option<String>

The engine type that is obtained from the server when the backup is created.

engine_model: Option<String>

The engine model that is obtained from the server when the backup is created.

engine_version: Option<String>

The engine version that is obtained from the server when the backup is created.

instance_profile_arn: Option<String>

The EC2 instance profile ARN that is obtained from the server when the backup is created. Because this value is stored, you are not required to provide the InstanceProfileArn again if you restore a backup.

instance_type: Option<String>

The instance type that is obtained from the server when the backup is created.

key_pair: Option<String>

The key pair that is obtained from the server when the backup is created.

preferred_backup_window: Option<String>

The preferred backup period that is obtained from the server when the backup is created.

preferred_maintenance_window: Option<String>

The preferred maintenance period that is obtained from the server when the backup is created.

s3_log_url: Option<String>

The Amazon S3 URL of the backup's log file.

security_group_ids: Option<Vec<String>>

The security group IDs that are obtained from the server when the backup is created.

server_name: Option<String>

The name of the server from which the backup was made.

service_role_arn: Option<String>

The service role ARN that is obtained from the server when the backup is created.

status: Option<String>

The status of a backup while in progress.

status_description: Option<String>

An informational message about backup status.

subnet_ids: Option<Vec<String>>

The subnet IDs that are obtained from the server when the backup is created.

tools_version: Option<String>

The version of AWS OpsWorks for Chef Automate-specific tools that is obtained from the server when the backup is created.

user_arn: Option<String>

The IAM user ARN of the requester for manual backups. This field is empty for automated backups.