Struct rusoto_opsworkscm::CreateServerRequest

pub struct CreateServerRequest {
    pub associate_public_ip_address: Option<bool>,
    pub backup_id: Option<String>,
    pub backup_retention_count: Option<i64>,
    pub disable_automated_backup: Option<bool>,
    pub engine: Option<String>,
    pub engine_attributes: Option<Vec<EngineAttribute>>,
    pub engine_model: Option<String>,
    pub engine_version: Option<String>,
    pub instance_profile_arn: String,
    pub instance_type: String,
    pub key_pair: Option<String>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub security_group_ids: Option<Vec<String>>,
    pub server_name: String,
    pub service_role_arn: String,
    pub subnet_ids: Option<Vec<String>>,
}

Fields
associate_public_ip_address: Option<bool>

Associate a public IP address with a server that you are launching. Valid values are true or false. The default value is true.

backup_id: Option<String>

If you specify this field, AWS OpsWorks for Chef Automate creates the server by using the backup represented by BackupId.

backup_retention_count: Option<i64>

The number of automated backups that you want to keep. Whenever a new backup is created, AWS OpsWorks for Chef Automate deletes the oldest backups if this number is exceeded. The default value is 1.

disable_automated_backup: Option<bool>

Enable or disable scheduled backups. Valid values are true or false. The default value is true.

engine: Option<String>

The configuration management engine to use. Valid values include Chef.

engine_attributes: Option<Vec<EngineAttribute>>

Optional engine attributes on a specified server.

Attributes accepted in a createServer request:

CHEF_PIVOTAL_KEY: A base64-encoded RSA private key that is not stored by AWS OpsWorks for Chef. This private key is required to access the Chef API. When no CHEF_PIVOTAL_KEY is set, one is generated and returned in the response.

CHEF_DELIVERY_ADMIN_PASSWORD: The password for the administrative user in the Chef Automate GUI. The password length is a minimum of eight characters, and a maximum of 32. The password can contain letters, numbers, and special characters (!/@#$%&+=_). The password must contain at least one lower case letter, one upper case letter, one number, and one special character. When no CHEF_DELIVERY_ADMIN_PASSWORD is set, one is generated and returned in the response.

engine_model: Option<String>

The engine model, or option. Valid values include Single.

engine_version: Option<String>

The major release version of the engine that you want to use. Values depend on the engine that you choose.

instance_profile_arn: String

The ARN of the instance profile that your Amazon EC2 instances use. Although the AWS OpsWorks console typically creates the instance profile for you, if you are using API commands instead, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the instance profile you need.

instance_type: String

The Amazon EC2 instance type to use. Valid values must be specified in the following format: ^([cm][34]|t2).* For example, m4.large. Valid values are t2.medium, m4.large, or m4.2xlarge.

key_pair: Option<String>

The Amazon EC2 key pair to set for the instance. This parameter is optional; if desired, you may specify this parameter to connect to your instances by using SSH.

preferred_backup_window: Option<String>

The start time for a one-hour period during which AWS OpsWorks for Chef Automate backs up application-level data on your server if automated backups are enabled. Valid values must be specified in one of the following formats:

HH:MM for daily backups

DDD:HH:MM for weekly backups

The specified time is in coordinated universal time (UTC). The default value is a random, daily start time.

Example: 08:00, which represents a daily start time of 08:00 UTC.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

preferred_maintenance_window: Option<String>

The start time for a one-hour period each week during which AWS OpsWorks for Chef Automate performs maintenance on the instance. Valid values must be specified in the following format: DDD:HH:MM. The specified time is in coordinated universal time (UTC). The default value is a random one-hour period on Tuesday, Wednesday, or Friday. See TimeWindowDefinition for more information.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

security_group_ids: Option<Vec<String>>

A list of security group IDs to attach to the Amazon EC2 instance. If you add this parameter, the specified security groups must be within the VPC that is specified by SubnetIds.

If you do not specify this parameter, AWS OpsWorks for Chef Automate creates one new security group that uses TCP ports 22 and 443, open to 0.0.0.0/0 (everyone).

server_name: String

The name of the server. The server name must be unique within your AWS account, within each region. Server names must start with a letter; then letters, numbers, or hyphens (-) are allowed, up to a maximum of 40 characters.

service_role_arn: String

The service role that the AWS OpsWorks for Chef Automate service backend uses to work with your account. Although the AWS OpsWorks management console typically creates the service role for you, if you are using the AWS CLI or API commands, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-stuff/latest/service-role-creation.yaml. This template creates a CloudFormation stack that includes the service role that you need.

subnet_ids: Option<Vec<String>>

The IDs of subnets in which to launch the server EC2 instance.

Amazon EC2-Classic customers: This field is required. All servers must run within a VPC. The VPC must have "Auto Assign Public IP" enabled.

EC2-VPC customers: This field is optional. If you do not specify subnet IDs, your EC2 instances are created in a default subnet that is selected by Amazon EC2. If you specify subnet IDs, the VPC must have "Auto Assign Public IP" enabled.

For more information about supported Amazon EC2 platforms, see Supported Platforms.