Struct rusoto_opsworkscm::OpsWorksCMClient

pub struct OpsWorksCMClient { /* fields omitted */ }

A client for the OpsWorksCM API.

Methods

impl OpsWorksCMClient

pub fn new(region: Region) -> OpsWorksCMClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> OpsWorksCMClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,