Struct rusoto_opsworkscm::Server

pub struct Server {
    pub associate_public_ip_address: Option<bool>,
    pub backup_retention_count: Option<i64>,
    pub cloud_formation_stack_arn: Option<String>,
    pub created_at: Option<f64>,
    pub disable_automated_backup: Option<bool>,
    pub endpoint: Option<String>,
    pub engine: Option<String>,
    pub engine_attributes: Option<Vec<EngineAttribute>>,
    pub engine_model: Option<String>,
    pub engine_version: Option<String>,
    pub instance_profile_arn: Option<String>,
    pub instance_type: Option<String>,
    pub key_pair: Option<String>,
    pub maintenance_status: Option<String>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub security_group_ids: Option<Vec<String>>,
    pub server_arn: Option<String>,
    pub server_name: Option<String>,
    pub service_role_arn: Option<String>,
    pub status: Option<String>,
    pub status_reason: Option<String>,
    pub subnet_ids: Option<Vec<String>>,
}

Describes a configuration management server.

Fields
associate_public_ip_address: Option<bool>

Associate a public IP address with a server that you are launching.

backup_retention_count: Option<i64>

The number of automated backups to keep.

cloud_formation_stack_arn: Option<String>

The ARN of the CloudFormation stack that was used to create the server.

created_at: Option<f64>

Time stamp of server creation. Example 2016-07-29T13:38:47.520Z

disable_automated_backup: Option<bool>

Disables automated backups. The number of stored backups is dependent on the value of PreferredBackupCount.

endpoint: Option<String>

A DNS name that can be used to access the engine. Example: myserver-asdfghjkl.us-east-1.opsworks.io

engine: Option<String>

The engine type of the server. The valid value in this release is Chef.

engine_attributes: Option<Vec<EngineAttribute>>

The response of a createServer() request returns the master credential to access the server in EngineAttributes. These credentials are not stored by AWS OpsWorks for Chef Automate; they are returned only as part of the result of createServer().

Attributes returned in a createServer response:

CHEF_PIVOTAL_KEY: A base64-encoded RSA private key that is generated by AWS OpsWorks for Chef Automate. This private key is required to access the Chef API.

CHEF_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Chef starter kit, which includes a README, a configuration file, and the required RSA private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents. From this directory, you can run Knife commands.

engine_model: Option<String>

The engine model of the server. The valid value in this release is Single.

engine_version: Option<String>

The engine version of the server. Because Chef is the engine available in this release, the valid value for EngineVersion is 12.

instance_profile_arn: Option<String>

The instance profile ARN of the server.

instance_type: Option<String>

The instance type for the server, as specified in the CloudFormation stack. This might not be the same instance type that is shown in the EC2 console.

key_pair: Option<String>

The key pair associated with the server.

maintenance_status: Option<String>

The status of the most recent server maintenance run. Shows SUCCESS or FAILED.

preferred_backup_window: Option<String>

The preferred backup period specified for the server.

preferred_maintenance_window: Option<String>

The preferred maintenance period specified for the server.

security_group_ids: Option<Vec<String>>

The security group IDs for the server, as specified in the CloudFormation stack. These might not be the same security groups that are shown in the EC2 console.

server_arn: Option<String>

The ARN of the server.

server_name: Option<String>

The name of the server.

service_role_arn: Option<String>

The service role ARN used to create the server.

status: Option<String>

The server's status. This field displays the states of actions in progress, such as creating, running, or backing up the server, as well as the server's health state.

status_reason: Option<String>

Depending on the server status, this field has either a human-readable message (such as a create or backup error), or an escaped block of JSON (used for health check results).

subnet_ids: Option<Vec<String>>

The subnet IDs specified in a CreateServer request.