Trait rusoto_opsworkscm::OpsWorksCM

pub trait OpsWorksCM {
    fn associate_node(
        &self, 
        input: AssociateNodeRequest
    ) -> RusotoFuture<AssociateNodeResponse, AssociateNodeError>;

    fn create_backup(
        &self, 
        input: CreateBackupRequest
    ) -> RusotoFuture<CreateBackupResponse, CreateBackupError>;

    fn create_server(
        &self, 
        input: CreateServerRequest
    ) -> RusotoFuture<CreateServerResponse, CreateServerError>;

    fn delete_backup(
        &self, 
        input: DeleteBackupRequest
    ) -> RusotoFuture<DeleteBackupResponse, DeleteBackupError>;

    fn delete_server(
        &self, 
        input: DeleteServerRequest
    ) -> RusotoFuture<DeleteServerResponse, DeleteServerError>;

    fn describe_account_attributes(
        &self
    ) -> RusotoFuture<DescribeAccountAttributesResponse, DescribeAccountAttributesError>;

    fn describe_backups(
        &self, 
        input: DescribeBackupsRequest
    ) -> RusotoFuture<DescribeBackupsResponse, DescribeBackupsError>;

    fn describe_events(
        &self, 
        input: DescribeEventsRequest
    ) -> RusotoFuture<DescribeEventsResponse, DescribeEventsError>;

    fn describe_node_association_status(
        &self, 
        input: DescribeNodeAssociationStatusRequest
    ) -> RusotoFuture<DescribeNodeAssociationStatusResponse, DescribeNodeAssociationStatusError>;

    fn describe_servers(
        &self, 
        input: DescribeServersRequest
    ) -> RusotoFuture<DescribeServersResponse, DescribeServersError>;

    fn disassociate_node(
        &self, 
        input: DisassociateNodeRequest
    ) -> RusotoFuture<DisassociateNodeResponse, DisassociateNodeError>;

    fn export_server_engine_attribute(
        &self, 
        input: ExportServerEngineAttributeRequest
    ) -> RusotoFuture<ExportServerEngineAttributeResponse, ExportServerEngineAttributeError>;

    fn restore_server(
        &self, 
        input: RestoreServerRequest
    ) -> RusotoFuture<RestoreServerResponse, RestoreServerError>;

    fn start_maintenance(
        &self, 
        input: StartMaintenanceRequest
    ) -> RusotoFuture<StartMaintenanceResponse, StartMaintenanceError>;

    fn update_server(
        &self, 
        input: UpdateServerRequest
    ) -> RusotoFuture<UpdateServerResponse, UpdateServerError>;

    fn update_server_engine_attributes(
        &self, 
        input: UpdateServerEngineAttributesRequest
    ) -> RusotoFuture<UpdateServerEngineAttributesResponse, UpdateServerEngineAttributesError>;
}

Trait representing the capabilities of the OpsWorksCM API. OpsWorksCM clients implement this trait.

Required methods

fn associate_node(
&self,
input: AssociateNodeRequest
) -> RusotoFuture<AssociateNodeResponse, AssociateNodeError>

Associates a new node with the server. For more information about how to disassociate a node, see DisassociateNode.

On a Chef server: This command is an alternative to knife bootstrap.

Example (Chef): aws opsworks-cm associate-node --server-name MyServer --node-name MyManagedNode --engine-attributes "Name=CHEF_ORGANIZATION,Value=default" "Name=CHEF_NODE_PUBLIC_KEY,Value=public-key-pem"

On a Puppet server, this command is an alternative to the puppet cert sign command that signs a Puppet node CSR.

Example (Chef): aws opsworks-cm associate-node --server-name MyServer --node-name MyManagedNode --engine-attributes "Name=PUPPET_NODE_CSR,Value=csr-pem"

A node can can only be associated with servers that are in a HEALTHY state. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid. The AssociateNode API call can be integrated into Auto Scaling configurations, AWS Cloudformation templates, or the user data of a server's instance.

fn create_backup(
&self,
input: CreateBackupRequest
) -> RusotoFuture<CreateBackupResponse, CreateBackupError>

Creates an application-level backup of a server. While the server is in the BACKING_UP state, the server cannot be changed, and no additional backup can be created.

Backups can be created for servers in RUNNING, HEALTHY, and UNHEALTHY states. By default, you can create a maximum of 50 manual backups.

This operation is asynchronous.

A LimitExceededException is thrown when the maximum number of manual backups is reached. An InvalidStateException is thrown when the server is not in any of the following states: RUNNING, HEALTHY, or UNHEALTHY. A ResourceNotFoundException is thrown when the server is not found. A ValidationException is thrown when parameters of the request are not valid.

fn create_server(
&self,
input: CreateServerRequest
) -> RusotoFuture<CreateServerResponse, CreateServerError>

Creates and immedately starts a new server. The server is ready to use when it is in the HEALTHY state. By default, you can create a maximum of 10 servers.

This operation is asynchronous.

A LimitExceededException is thrown when you have created the maximum number of servers (10). A ResourceAlreadyExistsException is thrown when a server with the same name already exists in the account. A ResourceNotFoundException is thrown when you specify a backup ID that is not valid or is for a backup that does not exist. A ValidationException is thrown when parameters of the request are not valid.

If you do not specify a security group by adding the SecurityGroupIds parameter, AWS OpsWorks creates a new security group.

Chef Automate: The default security group opens the Chef server to the world on TCP port 443. If a KeyName is present, AWS OpsWorks enables SSH access. SSH is also open to the world on TCP port 22.

Puppet Enterprise: The default security group opens TCP ports 22, 443, 4433, 8140, 8142, 8143, and 8170. If a KeyName is present, AWS OpsWorks enables SSH access. SSH is also open to the world on TCP port 22.

By default, your server is accessible from any IP address. We recommend that you update your security group rules to allow access from known IP addresses and address ranges only. To edit security group rules, open Security Groups in the navigation pane of the EC2 management console.

fn delete_backup(
&self,
input: DeleteBackupRequest
) -> RusotoFuture<DeleteBackupResponse, DeleteBackupError>

Deletes a backup. You can delete both manual and automated backups. This operation is asynchronous.

An InvalidStateException is thrown when a backup deletion is already in progress. A ResourceNotFoundException is thrown when the backup does not exist. A ValidationException is thrown when parameters of the request are not valid.

fn delete_server(
&self,
input: DeleteServerRequest
) -> RusotoFuture<DeleteServerResponse, DeleteServerError>

Deletes the server and the underlying AWS CloudFormation stacks (including the server's EC2 instance). When you run this command, the server state is updated to DELETING. After the server is deleted, it is no longer returned by DescribeServer requests. If the AWS CloudFormation stack cannot be deleted, the server cannot be deleted.

This operation is asynchronous.

An InvalidStateException is thrown when a server deletion is already in progress. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

fn describe_account_attributes(
&self
) -> RusotoFuture<DescribeAccountAttributesResponse, DescribeAccountAttributesError>

Describes your OpsWorks-CM account attributes.

This operation is synchronous.

fn describe_backups(
&self,
input: DescribeBackupsRequest
) -> RusotoFuture<DescribeBackupsResponse, DescribeBackupsError>

Describes backups. The results are ordered by time, with newest backups first. If you do not specify a BackupId or ServerName, the command returns all backups.

This operation is synchronous.

A ResourceNotFoundException is thrown when the backup does not exist. A ValidationException is raised when parameters of the request are not valid.

fn describe_events(
&self,
input: DescribeEventsRequest
) -> RusotoFuture<DescribeEventsResponse, DescribeEventsError>

Describes events for a specified server. Results are ordered by time, with newest events first.

This operation is synchronous.

A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

fn describe_node_association_status(
&self,
input: DescribeNodeAssociationStatusRequest
) -> RusotoFuture<DescribeNodeAssociationStatusResponse, DescribeNodeAssociationStatusError>

Returns the current status of an existing association or disassociation request.

A ResourceNotFoundException is thrown when no recent association or disassociation request with the specified token is found, or when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

fn describe_servers(
&self,
input: DescribeServersRequest
) -> RusotoFuture<DescribeServersResponse, DescribeServersError>

Lists all configuration management servers that are identified with your account. Only the stored results from Amazon DynamoDB are returned. AWS OpsWorks CM does not query other services.

This operation is synchronous.

A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

fn disassociate_node(
&self,
input: DisassociateNodeRequest
) -> RusotoFuture<DisassociateNodeResponse, DisassociateNodeError>

Disassociates a node from an AWS OpsWorks CM server, and removes the node from the server's managed nodes. After a node is disassociated, the node key pair is no longer valid for accessing the configuration manager's API. For more information about how to associate a node, see AssociateNode.

A node can can only be disassociated from a server that is in a HEALTHY state. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

fn export_server_engine_attribute(
&self,
input: ExportServerEngineAttributeRequest
) -> RusotoFuture<ExportServerEngineAttributeResponse, ExportServerEngineAttributeError>

Exports a specified server engine attribute as a base64-encoded string. For example, you can export user data that you can use in EC2 to associate nodes with a server.

This operation is synchronous.

A ValidationException is raised when parameters of the request are not valid. A ResourceNotFoundException is thrown when the server does not exist. An InvalidStateException is thrown when the server is in any of the following states: CREATING, TERMINATED, FAILED or DELETING.

fn restore_server(
&self,
input: RestoreServerRequest
) -> RusotoFuture<RestoreServerResponse, RestoreServerError>

Restores a backup to a server that is in a CONNECTION_LOST, HEALTHY, RUNNING, UNHEALTHY, or TERMINATED state. When you run RestoreServer, the server's EC2 instance is deleted, and a new EC2 instance is configured. RestoreServer maintains the existing server endpoint, so configuration management of the server's client devices (nodes) should continue to work.

This operation is asynchronous.

An InvalidStateException is thrown when the server is not in a valid state. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

fn start_maintenance(
&self,
input: StartMaintenanceRequest
) -> RusotoFuture<StartMaintenanceResponse, StartMaintenanceError>

Manually starts server maintenance. This command can be useful if an earlier maintenance attempt failed, and the underlying cause of maintenance failure has been resolved. The server is in an UNDER_MAINTENANCE state while maintenance is in progress.

Maintenance can only be started on servers in HEALTHY and UNHEALTHY states. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

fn update_server(
&self,
input: UpdateServerRequest
) -> RusotoFuture<UpdateServerResponse, UpdateServerError>

Updates settings for a server.

This operation is synchronous.

fn update_server_engine_attributes(
&self,
input: UpdateServerEngineAttributesRequest
) -> RusotoFuture<UpdateServerEngineAttributesResponse, UpdateServerEngineAttributesError>

Updates engine-specific attributes on a specified server. The server enters the MODIFYING state when this operation is in progress. Only one update can occur at a time. You can use this command to reset a Chef server's public key (CHEF_PIVOTAL_KEY) or a Puppet server's admin password (PUPPET_ADMIN_PASSWORD).

This operation is asynchronous.

This operation can only be called for servers in HEALTHY or UNHEALTHY states. Otherwise, an InvalidStateException is raised. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Loading content...